엔티티 오류 수정

+++ src/main/java/project/domain/repository/BoardNoticeRepository.java
@@ -6,8 +6,6 @@ import org.springframework.stereotype.Repository;
 import project.domain.entity.BoardNoticeEntity;
 
 @Repository
-public interface BoardNoticeRepository extends JpaRepository<BoardNoticeEntity, Integer>{
-
-	
+public interface BoardNoticeRepository extends JpaRepository<BoardNoticeEntity, Long>{
 
 }
